pyproject.toml
[project]
name = "djangoRestFramework-learn"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Foshati", email = "amirrezafoshati@gmail.com"},
]
dependencies = [
    "django>=5.0.7",
    "djangorestframework>=3.15.2",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true



[tool.pdm.scripts]
dev = "python manage.py runserver"
makemigrations = "python manage.py makemigrations"
migrate = "python manage.py migrate"
admin="python manage.py createsuperuser"

